remove jmessage api
