+++ README.md
@@ -229,58 +229,3 @@
             		LOG.info("Error Message: " + e.getErrorMessage());
 		}
 ```
-
-### IM User management 样例
-
-> 以下片断来自项目代码里的文件：example / cn.jpush.api.examples.IMUserExample
-
-```
-        JMessageClient client = new JMessageClient(appkey, masterSecret);
-        try {
-
-            List<RegisterInfo> users = new ArrayList<RegisterInfo>();
-
-            RegisterInfo user = RegisterInfo.newBuilder()
-                    .setUsername("test_user")
-                    .setPassword("test_pass")
-                    .build();
-
-            RegisterInfo user1 = RegisterInfo.newBuilder()
-                    .setUsername("test_user1")
-                    .setPassword("test_pass1")
-                    .build();
-
-            users.add(user);
-            users.add(user1);
-
-            RegisterInfo[] regUsers = new RegisterInfo[users.size()];
-
-            String res = client.registerUsers(users.toArray(regUsers));
-            LOG.info(res);
-        } catch (APIConnectionException e) {
-            LOG.error("Connection error. Should retry later. ", e);
-        } catch (APIRequestException e) {
-            LOG.error("Error response from JPush server. Should review and fix it. ", e);
-            LOG.info("HTTP Status: " + e.getStatus());
-            LOG.info("Error Message: " + e.getMessage());
-        }
-```
-
-### IM Group management 样例
-
-> 以下片断来自项目代码里的文件：example / cn.jpush.api.examples.IMGroupExample
-
-```
-        JMessageClient client = new JMessageClient(appkey, masterSecret);
-        try {
-            String res = client.createGroup("test_user", "test_gname1", "description", "test_user");
-            LOG.info(res);
-        } catch (APIConnectionException e) {
-            LOG.error("Connection error. Should retry later. ", e);
-        } catch (APIRequestException e) {
-            LOG.error("Error response from JPush server. Should review and fix it. ", e);
-            LOG.info("HTTP Status: " + e.getStatus());
-            LOG.info("Error Message: " + e.getMessage());
-        }
-```
-
