--- src/main/java/cn/jpush/api/jmessage/base/ServiceConstant.java
@@ -1,17 +1,0 @@
-package cn.jpush.api.jmessage.base;
-
-public class ServiceConstant {
-
-    public static final String HOST_NAME_SSL = "https://api.im.jpush.cn";
-
-    public static final String ADMIN_PATH = "/admins";
-
-    public static final String USER_PATH = "/users";
-
-    public static final String GROUP_PATH = "/groups";
-
-    public static final String CONTACTS_PATH = "/";
-
-    public static final String MESSAGE_PATH = "/";
-
-}
