--- src/main/java/cn/jpush/api/jmessage/base/model/IModel.java
@@ -1,9 +1,0 @@
-package cn.jpush.api.jmessage.base.model;
-
-import com.google.gson.JsonElement;
-
-public interface IModel {
-
-    public JsonElement toJSON();
-
-}
