--- src/main/java/cn/jpush/api/jmessage/contacts/ContactsClient.java
@@ -1,8 +1,0 @@
-package cn.jpush.api.jmessage.contacts;
-
-/**
- *
- * Not yet opened.
- */
-public class ContactsClient {
-}
