--- src/main/java/cn/jpush/api/jmessage/contacts/package-info.java
@@ -1,6 +1,0 @@
-/**
- * Contacts management API features.
- *
- * Not yet
- */
-package cn.jpush.api.jmessage.contacts;
