--- src/main/java/cn/jpush/api/jmessage/group/package-info.java
@@ -1,6 +1,0 @@
-/**
- * Group management API features.
- *
- * Url: https://api.im.jpush.cn/groups
- */
-package cn.jpush.api.jmessage.group;
