--- src/main/java/cn/jpush/api/jmessage/message/MessageClient.java
@@ -1,8 +1,0 @@
-package cn.jpush.api.jmessage.message;
-
-/**
- *
- * Not yet opened.
- */
-public class MessageClient {
-}
