--- src/main/java/cn/jpush/api/jmessage/message/package-info.java
@@ -1,6 +1,0 @@
-/**
- * Message management API features.
- *
- * Not yet.
- */
-package cn.jpush.api.jmessage.message;
