--- src/main/java/cn/jpush/api/jmessage/package-info.java
@@ -1,6 +1,0 @@
-/**
- * JPush IM (JMessage) API features.
- *
- * Url: https://api.im.jpush.cn
- */
-package cn.jpush.api.jmessage;
