--- src/main/java/cn/jpush/api/jmessage/user/package-info.java
@@ -1,6 +1,0 @@
-/**
- * User management API features.
- *
- * Url: https://api.im.jpush.cn/users
- */
-package cn.jpush.api.jmessage.user;
