+++ src/main/resources/javadoc-overview.html
@@ -5,6 +5,6 @@
 <body>
     JPush API Java 客户端。
 
-    主要分为 3 个功能部分：推送、报表、IM管理
+    主要分为 2 个功能部分：推送、报表
 </body>
 </html>
